Prefer Logger.warning()

+++ server/worker_srv.py
@@ -72,7 +72,7 @@ class WorkerManager:
 		if pipeline_def.config is not None:
 			# Inline config
 			if pipeline_def.configPath is not None:
-				self.log.warn("Camera %s requested pipeline %s with both an internal and external config", cid, pipeline_id)
+				self.log.warning("Camera %s requested pipeline %s with both an internal and external config", cid, pipeline_id)
 			nn_cfg = pipeline_def.config
 		elif pipeline_def.configPath is not None:
 			# Load pipeline from file
@@ -123,7 +123,7 @@ class WorkerManager:
 				return cached
 			apriltagPath = self._resolve_path(apriltag.path)
 			if not apriltagPath.exists():
-				self.log.warn("Camera %s requested SLAM with AprilTags at %s, but that file doesn't exist", cid, apriltagPath)
+				self.log.warning("Camera %s requested SLAM with AprilTags at %s, but that file doesn't exist", cid, apriltagPath)
 				return None
 			
 			# Load FRC data
@@ -181,7 +181,7 @@ class WorkerManager:
 				return cached
 			apriltagPath = self._resolve_path(apriltag.path)
 			if not apriltagPath.exists():
-				self.log.warn("Camera %s requested SLAM with AprilTags at %s, but that file doesn't exist", cid, apriltagPath)
+				self.log.warning("Camera %s requested SLAM with AprilTags at %s, but that file doesn't exist", cid, apriltagPath)
 				return None
 			# Check that it's correct
 			with open(apriltagPath, 'r') as f:
@@ -224,7 +224,7 @@ class WorkerManager:
 		elif raw_slam == True:
 			slam_cfg: SlamConfig = self.config.slam
 			if slam_cfg is None:
-				self.log.warn("Camera %s requested global SLAM, but no config exists", cid)
+				self.log.warning("Camera %s requested global SLAM, but no config exists", cid)
 				return None
 		else:
 			slam_cfg = raw_slam
@@ -233,7 +233,7 @@ class WorkerManager:
 		try:
 			apriltagPath = self._resolve_apriltag(cid, slam_cfg.apriltag)
 			slam_cfg: SlamConfigBase
-			return worker.SlamConfig(
+			return worker.WorkerSlamConfig(
 				backend=slam_cfg.backend,
 				syncNN=slam_cfg.syncNN,
 				slam=slam_cfg.slam,
@@ -244,18 +244,18 @@ class WorkerManager:
 			self.log.exception("Error resolving SLAM config for camera %s", cid)
 			return None
 
-	def process_one(self, camera: CameraConfig, idx: int = 0) -> worker.InitConfig:
+	def process_one(self, camera: CameraConfig, idx: int = 0) -> worker.WorkerInitConfig:
 		"Resolve the config for a single camera"
 		cid = CameraId(idx, camera.id)
 		selector = self._resolve_selector(cid, camera.selector)
 		slam_cfg = self._resolve_slam(cid, camera.slam)
 		object_detection = self._resolve_pipeline(cid, camera.object_detection)
 
-		return worker.InitConfig(
+		return worker.WorkerInitConfig(
 			id=camera.id,
 			selector=selector,
 			max_usb=camera.max_usb,
-			optional=camera.optional,
+			retry=camera.retry,
 			robot_to_camera=camera.pose,
 			slam=slam_cfg,
 			object_detection=object_detection,
@@ -265,8 +265,8 @@ class WorkerManager:
 		"Start all camera processes"
 		for i, camera in enumerate(self.config.cameras):
 			init_cfg = self.process_one(camera, i)
-			
-			wh = WorkerHandle(f"worker_{i}", init_cfg, log=self.log)
+			name = init_cfg.id if init_cfg.id is not None else f'cam_{i}'
+			wh = WorkerHandle(name, init_cfg, log=self.log)
 			self._workers.append(wh)
 			wh.start()
 	
@@ -292,50 +292,61 @@ class WorkerHandle:
 	"Manage a single camera worker"
 	
 	proc: Process
-	def __init__(self, name: str, config: worker.InitConfig, *, log: Optional[logging.Logger] = None, ctx: Optional['BaseContext'] = None) -> None:
-		from worker import main as worker_main
+	def __init__(self, name: str, config: worker.WorkerInitConfig, *, log: Optional[logging.Logger] = None, ctx: Optional['BaseContext'] = None):
 		if ctx is None:
 			ctx = get_context('spawn')
+		self._ctx = ctx
 		
 		self.name = name
 		self.child_state = None
 		self.log = log.getChild(name) if (log is not None) else logging.getLogger(name)
 		
 		self.config = config
-		print("worker config", config.model_dump_json())
+		self.log.info("Start with config %s", config.model_dump_json())
+		self.proc = None
 		self._restarts = 0
 		self.robot_to_camera = config.robot_to_camera
 		self.cmd_queue = ctx.Queue()
 		self.data_queue = ctx.Queue()
-		self.proc = ctx.Process(
-			target=worker_main,
-			args=[config, self.data_queue, self.cmd_queue],
-			daemon=True,
-			name=f'moenet_{name}'
-		)
 		self._require_flush_id = 0
 		self._last_flush_id = 0
 	
 	def start(self):
+		if self.proc is not None:
+			self.log.warning('Started twice!')
+			return
+		
+		from worker import main as worker_main
+		self.proc = self._ctx.Process(
+			target=worker_main,
+			args=[self.config, self.data_queue, self.cmd_queue],
+			daemon=True,
+			name=f'moenet_{self.name}'
+		)
 		self.proc.start()
 
 	def stop(self, send_command: bool = True):
 		if self.proc is None:
 			return
 		try:
+			# Try stopping nicely
 			try:
 				self.log.info("Stopping...")
 				if send_command:
 					self.cmd_queue.put(worker.CmdChangeState(target=worker.WorkerState.STOPPED), block=True, timeout=1.0)
-				self.proc.join(3.0)
+				self.proc.join(1.0)
 			except:
 				self.log.exception("Error on join")
+			
+			if self.proc.is_alive():
 				self.proc.kill()
-				try:
-					self.proc.join()
-				except:
-					self.log.exception('Worker exception on join')
+			
+			try:
+				self.proc.join()
+			except:
+				self.log.exception('Exception on join')
 		finally:
+			self.log.debug("close")
 			self.proc.close()
 		self.proc = None
 		self.log.info("Stopped")
@@ -347,7 +358,31 @@ class WorkerHandle:
 		self._require_flush_id += 1
 		self.send(worker.CmdFlush(id=self._require_flush_id))
 	
+	def _handle_dead(self):
+		optional = self.config.retry.optional
+		self._restarts += 1
+		can_retry = self._restarts < self.config.retry.connection_tries
+
+		if optional:
+			self.log.warning('Unexpectedly exited')
+			self.child_state = worker.WorkerState.STOPPING
+		else:
+			self.log.error('Unexpectedly exited')
+		# Cleanup process
+		self.stop(False)
+		if can_retry:
+			self.child_state = worker.WorkerState.STOPPED
+			self.log.info("Restarting (%d of %d)", self._restarts, self.config.retry.connection_tries)
+			# TODO: honor connection_delay?
+			self.start()
+		elif optional:
+			self.child_state = worker.WorkerState.STOPPED
+		else:
+			self.child_state = worker.WorkerState.FAILED
+			raise RuntimeError(f'Camera {self.name} unexpectedly exited')
+
 	def poll(self):
+		"Process messages from worker"
 		if self.proc is None:
 			return
 		
@@ -357,33 +392,26 @@ class WorkerHandle:
 			try:
 				# We want to process any remaining packets even if the process died
 				# but we know we don't need to block on the queue then
-				packet: Union[worker.MsgPose, worker.MsgDetections, worker.MsgChangeState] = self.data_queue.get(block=is_alive, timeout=0.01)
+				packet: worker.AnyMsg = self.data_queue.get(block=is_alive, timeout=0.01)
 			except Empty:
 				break
 			else:
 				if isinstance(packet, worker.MsgChangeState):
 					self.child_state = packet.current
 				elif isinstance(packet, worker.MsgFlush):
-					self.log.debug('Finished flush %d', packet.flush_id)
-					self._last_flush_id = max(self._last_flush_id, packet.flush_id)
+					self.log.debug('Finished flush %d', packet.id)
+					self._last_flush_id = max(self._last_flush_id, packet.id)
 					continue # We don't need to forward this
+				elif isinstance(packet, worker.MsgLog):
+					self.log.log(packet.level, packet.msg)
+					continue
 				
 				if self._last_flush_id < self._require_flush_id:
 					# Packets are invalidated by a flush
+					self.log.info("Skipping packet (flushed)")
 					continue
 
 				yield packet
 		
 		if not is_alive:
-			if not self.config.optional:
-				self.log.error('Camera process unexpectedly exited')
-			else:
-				self.log.warn('Camera process unexpectedly exited')
-			# Cleanup process
-			self.child_state = worker.WorkerState.FAILED
-			self.stop(False)
-
-			if (not self.config.optional):
-				raise RuntimeError('Camera unexpectedly exited')
-			
-			#TODO: restart-retry logic
+			self._handle_dead()
